implement strtotime() via timelib to make it identical to PHP strtotime()
PHP uses timelib more intensively and we could replace some of out
lookup tables from datetime.cpp with its help, but it might be a better
idea to do this (if ever) later.

This patch also includes a fix to date_default_timezone_get(),
so if anyone uses a date_default_timezone_set() with a supported
argument, calling date_default_timezone_get() would return that value.